Relax dependecy requirements to allow use with Rack 3
We need to upgrade to `rack` 3.x in our project. To do this we need to upgrade to `wasabi` to 5.x and `httpi` to 4.x, but this is prevented by the dependency requirements of `google-ads-common` and `google-ads-savon`.

According to the changelog, these versions are restricted to "tighten dependency requirements to maintain 2.7 support", but this makes no sense. A project is of course free to use older versions of these dependencies compatible with Ruby 2.7, but it should not be a requirement to use outdated gems if you are using a newer version of Ruby.

The newer versions of `akamai`, `faraday`, `faraday-net_http`, `httpi`, `nokogiri`, `nori`, `public_suffix`, `rack`, `rake`, and `wasabi` all seem to be compatible with the current implementation of `google-ads-common` and `google-ads-savon`.